Harden against Plugin Check (PCP): clear all actionable findings
Ran the official Plugin Check (the wordpress.org review suite) and resolved
every actionable error. Remaining PCP output is the deferred WP.org-only set:
the text-domain↔slug mismatch (~1387, pending a slug decision), the missing
readme.txt headers, and our intentional dotfiles.

- Direct-access guards: add `defined( 'ABSPATH' ) || exit;` to 27 PHP files
  that lacked it (block renders, abilities, REST controllers, etc.).
- External resource: stop loading select2 from a CDN (disallowed). Use
  WooCommerce's bundled select2/selectWoo (the plugin requires WC), with a
  JS guard so native selects degrade gracefully if it's absent.
- Output escaping (~102): wrap plain scalars in esc_html/esc_attr/(int); add
  justified phpcs:ignore for genuinely-safe markup (block wrapper attrs,
  interactivity context, paginate_links, esc_*-assembled HTML) — the code was
  already XSS-audited. Fixed two real `wp_die( __() )` → `esc_html__()`.
- Direct DB (~18): justified phpcs:ignore — every dynamic value is already
  $wpdb->prepare()'d; only $wpdb->prefix table names are interpolated.
- File ops: gmdate() for date(); wp_delete_file() for unlink(); justified
  ignores for legitimate php://output / php://temp / uploaded-CSV streams.
- Misc: convert two heredocs to plain strings; make a few i18n strings literal
  (or justified-ignore config-sourced labels); one translators comment.

Test harness: the new ABSPATH guard in helpers.php (loaded via composer
"files" at PHPUnit startup, before any bootstrap) was exiting all suites.
Drop helpers.php from composer autoload "files" and require it explicitly in
the unit bootstrap after defining ABSPATH; WP defines it for the other suites.

Unit 55, integration 88/277, wc 15/43; lint clean; PCP actionable findings: 0.

+++ includes/admin/class-menu.php
@@ -107,7 +107,7 @@ class Menu {
                             printf(
                                 /* translators: %d: donation count */
                                 esc_html__( '%d donations', 'starter-shelter' ),
-                                $stats['donations']['count'] ?? 0
+                                (int) ( $stats['donations']['count'] ?? 0 )
                             );
                             ?>
                         </p>
@@ -125,7 +125,7 @@ class Menu {
                             printf(
                                 /* translators: %d: number of new records this month */
                                 esc_html__( '%d new this month', 'starter-shelter' ),
-                                $stats['memberships']['new'] ?? 0
+                                (int) ( $stats['memberships']['new'] ?? 0 )
                             );
                             ?>
                         </p>
@@ -155,7 +155,7 @@ class Menu {
                             printf(
                                 /* translators: %d: number of new records this month */
                                 esc_html__( '%d new this month', 'starter-shelter' ),
-                                $stats['donors']['new'] ?? 0
+                                (int) ( $stats['donors']['new'] ?? 0 )
                             );
                             ?>
                         </p>
